move damage calculations to flood risk module
